refactor: separate clientversion from version
