+++ src/addrdb.cpp
@@ -7,7 +7,7 @@
 
 #include <addrman.h>
 #include <chainparams.h>
-// #include <clientversion.h>
+#include <clientversion.h>
 #include <hash.h>
 // #include <random.h>
 // #include <streams.h>
