+++ src/node/blockstorage.cpp
@@ -4,6 +4,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "chainparams.h"
+#include "clientversion.h"
 #include "main.h"
 #include "protocol.h"
 #include "serialize.h"
