+++ src/qt/forms/rpcconsole.ui
@@ -107,13 +107,6 @@
          </property>
         </widget>
        </item>
-       <item row="8" column="0">
-        <widget class="QLabel" name="buildDateLabel">
-         <property name="text">
-          <string>Build date</string>
-         </property>
-        </widget>
-       </item>
        <item row="47" column="0">
         <spacer name="verticalSpacer_2">
          <property name="orientation">
@@ -163,22 +156,6 @@
         <widget class="QLabel" name="clientVersionLabel">
          <property name="text">
           <string>Client version</string>
-         </property>
-        </widget>
-       </item>
-       <item row="8" column="2">
-        <widget class="QLabel" name="buildDate">
-         <property name="cursor">
-          <cursorShape>IBeamCursor</cursorShape>
-         </property>
-         <property name="text">
-          <string>N/A</string>
-         </property>
-         <property name="textFormat">
-          <enum>Qt::PlainText</enum>
-         </property>
-         <property name="textInteractionFlags">
-          <set>Qt::LinksAccessibleByMouse|Qt::TextSelectableByMouse</set>
          </property>
         </widget>
        </item>
