+++ src/qt/res/gridcoinresearch.rc
@@ -2,7 +2,7 @@ IDI_ICON1 ICON DISCARDABLE "icons/gridcoin.ico"
 IDI_ICON2 ICON DISCARDABLE "icons/gridcoin_testnet.ico"
 
 #include <windows.h>             // needed for VERSIONINFO
-#include "../../version.h" // holds the needed client version information
+#include "../../clientversion.h" // holds the needed client version information
 
 #define VER_PRODUCTVERSION     CLIENT_VERSION_MAJOR,CLIENT_VERSION_MINOR,CLIENT_VERSION_REVISION,CLIENT_VERSION_BUILD
 #define VER_PRODUCTVERSION_STR STRINGIZE(CLIENT_VERSION_MAJOR) "." STRINGIZE(CLIENT_VERSION_MINOR) "." STRINGIZE(CLIENT_VERSION_REVISION) "." STRINGIZE(CLIENT_VERSION_BUILD)
