+++ src/util.cpp
@@ -13,8 +13,6 @@
 #include <util/strencodings.h>
 #include <util/string.h>
 
-#include <boost/algorithm/string/join.hpp>
-#include <boost/algorithm/string/predicate.hpp> // for startswith() and endswith()
 #include <boost/date_time/posix_time/posix_time.hpp>  //For day of year
 #include <cmath>
 #include <boost/lexical_cast.hpp>
@@ -485,25 +483,6 @@ void seed_insecure_rand(bool fDeterministic)
     }
 }
 
-string FormatVersion(int nVersion)
-{
-    if (nVersion%100 == 0)
-        return strprintf("%d.%d.%d", nVersion/1000000, (nVersion/10000)%100, (nVersion/100)%100);
-    else
-        return strprintf("%d.%d.%d.%d", nVersion/1000000, (nVersion/10000)%100, (nVersion/100)%100, nVersion%100);
-}
-
-#ifndef UPGRADERFLAG
-// avoid including unnecessary files for standalone upgrader
-
-
-string FormatFullVersion()
-{
-    return CLIENT_BUILD;
-}
-
-#endif
-
 double Round(double d, int place)
 {
     const double accuracy = std::pow(10, place);
@@ -551,19 +530,6 @@ std::vector<std::string> split(const std::string& s, const std::string& delim)
     // Append final value
     elems.push_back(s.substr(pos, end - pos));
     return elems;
-}
-
-// Format the subversion field according to BIP 14 spec (https://en.bitcoin.it/wiki/BIP_0014)
-std::string FormatSubVersion(const std::string& name, int nClientVersion, const std::vector<std::string>& comments)
-{
-    std::ostringstream ss;
-    ss << "/";
-    ss << name << ":" << FormatVersion(nClientVersion);
-
-    if (!comments.empty())         ss << "(" << boost::algorithm::join(comments, "; ") << ")";
-
-    ss << "/";
-    return ss.str();
 }
 
 void runCommand(std::string strCommand)
