+++ src/wallet/db.h
@@ -6,10 +6,10 @@
 #ifndef BITCOIN_WALLET_DB_H
 #define BITCOIN_WALLET_DB_H
 
+#include "clientversion.h"
 #include "fs.h"
 #include "streams.h"
 #include "sync.h"
-#include "version.h"
 
 #include <map>
 #include <string>
